Moved save functionality to database.js 	modified:   config/routes/database.js 	modified:   index.js
